demo(dashboard): extract SCREENPLAY parser into screenplay.ts
Behavior-preserving refactor: move the SCREENPLAY.md parser + typed model
(Segment/Beat/Demo, parseBeats/parseGapWow/findResetCmd, esc/inlineMd) out of
build.ts into demo/dashboard/screenplay.ts. build.ts keeps the renderer,
HTML template, and evidence/timeline code.

control.html output is byte-identical: a fresh build from HEAD and from the
refactored code both produce sha 3b3690bd. Foundation for the live-control
Vite/React rebuild — the typed model is the seam a React consumer reads.
